requirements.txt
boto3==1.34.144
moto[s3]==5.0.0
kafka-python==2.0.2
click==8.1.7
